Extract raw query functionality into cosmwasm
